Add files via upload
Vehicle spawning algorithm.  Implemented the spawning vehicles with the randomity of (1/1000), the vehicle ID.  when spawning the direction of spawn, and destination should be pre-determined, and all should be random. created the all possibilities in the switch case and randomized it. All the vehicles are spawning on the right side of the road. 
references: https://cplusplus.com/reference/cstdlib/rand/